aider: Add the Bitwarden CLI Node app to the `vm-setup/nodejs` role and update the default package list with the latest version.
# Aider chat conversation:

USER: Add the bitwarden cli node app to the vm-setup/nodejs role

ASSISTANT: To add the Bitwarden CLI Node app to the `vm-setup/nodejs` role, we need to modify the `tasks/main.yml` file within that role to include a task that installs the Bitwarden CLI using npm (Node Package Manager).

Here is the diff for the required changes:

```diff
--- vm-setup/roles/nodejs/tasks/main.yml
+++ vm-setup/roles/nodejs/tasks/main.yml
@@ ... @@
 # Ensure Node.js and npm are installed (using Node Version Manager)
 - name: Install NVM (Node Version Manager)
   ...

+# Install Bitwarden CLI using npm
+- name: Install Bitwarden CLI
+  npm:
+    name: "@bitwarden/cli"
+    global: yes
+    state: present
+
 # Additional tasks can be placed here
 ...
```

This task will install the Bitwarden CLI globally on the system using npm, which should be available after the Node.js installation tasks that are presumably already defined in the `vm-setup/roles/nodejs/tasks/main.yml` file.

USER: The package should be added to the packages list in the defaults main.yml

ASSISTANT: To add the Bitwarden CLI Node app to the `nodejs` role's default package list, we need to modify the `defaults/main.yml` file within the `vm-setup/roles/nodejs` directory. Here is the diff for the required changes:

```diff
--- vm-setup/roles/nodejs/defaults/main.yml
+++ vm-setup/roles/nodejs/defaults/main.yml
@@ ... @@
 nodejs_packages:
   - name: "some-node-package"
     version: "1.0.0"
+  - name: "@bitwarden/cli"
+    version: "latest"
 # You can specify the version with the npm package name using '@'
 # For example: "package-name@1.0.0"
 ...
```

This change adds the Bitwarden CLI package to the list of Node.js packages to be installed, with the version set to "latest". The actual installation of these packages would be handled by the tasks in `tasks/main.yml`, which would loop over this list and ensure each package is installed.

USER: How do I find out what login gui I'm using (I'm using gnome).

ASSISTANT: To find out what login GUI you are using, you can check the display manager service that is active on your system. Since you are using GNOME, the default display manager is typically GDM (GNOME Display Manager). You can check the status of GDM or find out which display manager is running with the following command:

```sh
systemctl status display-manager.service
```

Alternatively, you can use the `dm-tool` command to get information about the current display manager session:

```sh
dm-tool info
```

These commands should be run in the terminal. If you need to run these commands as part of a script or automation, you can create a shell script and execute it. If you need assistance with creating such a script, please let me know, and I can provide you with the necessary code.

USER: Add task or tasks to vm-setup/roles/gnome-settings/tasks/gdm.yml that sets the background color of the login screen to be the same as the desktop background color.

ASSISTANT: To set the background color of the GDM (GNOME Display Manager) login screen to be the same as the desktop background color, we need to add tasks to the `vm-setup/roles/gnome-settings/tasks/gdm.yml` file. These tasks will use `gsettings` to set the background color for the GDM user.

Here is the diff for the required changes:

```diff
--- vm-setup/roles/gnome-settings/tasks/gdm.yml
+++ vm-setup/roles/gnome-settings/tasks/gdm.yml
@@ ... @@
 # This file should contain tasks specific to configuring GDM (GNOME Display Manager)

+# Get the current desktop background color
+- name: Get desktop background color
+  become: yes
+  command: gsettings get org.gnome.desktop.background primary-color
+  register: desktop_background_color
+  changed_when: false
+
+# Set GDM login screen background color to match desktop background
+- name: Set GDM login screen background color
+  become: yes
+  command: gsettings set org.gnome.desktop.background primary-color "{{ desktop_background_color.stdout }}"
+  become_user: gdm
+  when: desktop_background_color.stdout != ""
+
 ...
```

These tasks will first retrieve the current desktop background color and then set the GDM login screen background color to match it. The `gsettings` command needs to be run as the `gdm` user because GDM's settings are specific to that user. The `changed_when: false` line ensures that the task of getting the desktop background color does not report a change in the Ansible output, as it is only gathering information.