aider: Added task to set GDM login screen background color using primary color key from dictionary.
# Aider chat conversation:

USER: That variable is a dict, that key is primary-color.

ASSISTANT: Given that the `gnome_settings_desktop_background` variable is a dictionary and the key for the primary color is `primary-color`, I will provide the updated diff for the `vm-setup/roles/gnome_settings/tasks/gdm.yml` file to correctly set the GDM login screen background color using this key.

Here is the diff:

```diff
--- vm-setup/roles/gnome_settings/tasks/gdm.yml
+++ vm-setup/roles/gnome_settings/tasks/gdm.yml
@@ ... @@
+
+- name: Set GDM login screen background color
+  become: true
+  command: gsettings set org.gnome.login-screen background-color "{{ gnome_settings_desktop_background['primary-color'] }}"
+  become_user: gdm
+  when: gnome_settings_desktop_background['primary-color'] is defined
```

This diff adds a new task to the `gdm.yml` file that sets the GDM login screen background color to the `primary-color` value specified in the `gnome_settings_desktop_background` dictionary. The task uses `become: true` to gain superuser privileges and `become_user: gdm` to run the command as the GDM user. The `when` condition ensures that the task is only executed if the `primary-color` key is defined in the `gnome_settings_desktop_background` dictionary.