aider: Modify the `tilix` include task to only be used if `tilix_profile_filename` is set.
# Aider chat conversation:

USER: Modify the tilix include task to only be used if tilix_profile_filename is set

ASSISTANT: To modify the `tilix` include task in the `vm-setup/roles/gnome_settings/tasks/main.yml` file so that it only runs if `tilix_profile_filename` is set, I will add a conditional to the task. Here is the diff:

```diff
--- vm-setup/roles/gnome_settings/tasks/main.yml
+++ vm-setup/roles/gnome_settings/tasks/main.yml
@@ ... @@

-- name: Include Tilix handling task
-  include_tasks: tilix.yml
-  vars:
-    tilix_profile_filename: "{{ gnome_settings_tilix_profile_filename }}"
+- name: Include Tilix handling task if tilix_profile_filename is set
+  include_tasks: tilix.yml
+  when: gnome_settings_tilix_profile_filename is defined and gnome_settings_tilix_profile_filename | length > 0
```

This diff modifies the existing task by adding a `when` condition that checks if `gnome_settings_tilix_profile_filename` is defined and not empty before including the `tilix.yml` tasks. If `gnome_settings_tilix_profile_filename` is not set, the `tilix.yml` tasks will be skipped.