aider: Added task to add Tilix to the favorites menu bar if not already present.
# Aider chat conversation:

USER: In the tilix.yml file, add a task that adds the tilix app to the favorites menu bar if it's not already there.

ASSISTANT: ```diff
--- vm-setup/roles/gnome_settings/tasks/tilix.yml
+++ vm-setup/roles/gnome_settings/tasks/tilix.yml
@@ ... @@
   when: dconf_load_result is changed

+# Add Tilix to the favorites menu bar if it's not already there
+- name: Add Tilix to the favorites menu bar
+  block:
+    - name: Get current list of favorite apps
+      command: gsettings get org.gnome.shell favorite-apps
+      register: current_favorites
+
+    - name: Add Tilix to the list of favorite apps
+      command: gsettings set org.gnome.shell favorite-apps "{{ current_favorites.stdout | regex_replace(']', ', ') }}'tilix.desktop']"
+      when: "'tilix.desktop' not in current_favorites.stdout"
+  become: yes
+  become_user: "{{ ansible_env.SUDO_USER | default(ansible_env.USER) }}"
```
This diff adds a new task to the `tilix.yml` file that ensures Tilix is added to the GNOME favorites menu bar. It first retrieves the current list of favorite apps and then adds Tilix to the list if it's not already present. The task uses `become` to execute the command as the user who invoked sudo, falling back to the current user if `SUDO_USER` is not set.