aider: Add condition to exclude `version` subcommand from completion if `-version` or `version` is used.
# Aider chat conversation:

USER: -version is an alias for version, does that make a difference?

ASSISTANT: The fact that `-version` is an alias for `version` in Terraform does not require a change in the bash completion script. The completion script already handles `-version` as a special option, and since `version` is a subcommand, it will be included in the list of subcommands that are completed after `terraform`.

However, if you want to ensure that the `version` subcommand is not offered as a completion after `-version` is used, or vice versa, we can add a condition to handle this. Here's the diff for the necessary changes:

```diff
--- config/shell-startup/terraform
+++ config/shell-startup/terraform
@@ ... @@
 _terraform() {
   local cur="${COMP_WORDS[COMP_CWORD]}"
   local prev="${COMP_WORDS[COMP_CWORD-1]}"
   local special_opts="-help -version"
   local initial_opts="-chdir"
+  local version_used=""

+  # Check if -version or version has been used
+  if [[ " ${COMP_WORDS[*]} " =~ " -version " ]] || [[ " ${COMP_WORDS[*]} " =~ " version " ]]; then
+    version_used="yes"
+  fi

   if [[ " ${COMP_WORDS[*]} " =~ " -help " ]] || [[ " ${COMP_WORDS[*]} " =~ " -version " ]]; then
     COMPREPLY=()
   elif [[ "$prev" == "terraform" ]] && ! [[ " ${COMP_WORDS[*]:1} " =~ " -chdir " ]]; then
     local subcommands="$(terraform | awk '/^  [a-z-]/ {print $1}')"
+    # Remove version from subcommands if -version or version is used
+    if [[ "$version_used" == "yes" ]]; then
+      subcommands=$(echo "$subcommands" | tr ' ' '\n' | grep -v '^version$' | tr '\n' ' ')
+    fi
     COMPREPLY=($(compgen -W "${subcommands} ${special_opts} ${initial_opts}" -- "$cur"))
   elif [[ "$prev" == "-chdir" ]]; then
     COMPREPLY=($(compgen -d -- "$cur"))
   fi
 }

 complete -F _terraform terraform
```

This change adds a check to see if `-version` or `version` has been used anywhere in the command line. If so, it removes `version` from the list of subcommands to ensure it's not offered as a completion.